+++ RELEASE-NOTES.md
@@ -25,6 +25,7 @@ as release notes for end user on a Recalbox upgrade.
 - Add mc10 and coco3 model to trs80 color computer xroar emulator
 - Add CRT check menu footer
 - Display game regions flag in gamelist
+- Make default no image in gamelist themable
 
 ### Improvements